Add files via upload
First I declared a variable called res which will hold the maximum water value. After that I declared the for loop for every element of an array except first and the last element. I used the two pointer concept here i.e left and the right to get the maximum water value. After that I found the maximum element using max function on the left and then I compared it with the right and updated the res value. So here, we got the maximum water value.